paper_code_submission/language_detection/requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
cymem==2.0.6
Cython==0.29.28
dartsclone==0.9.0
DAWG-Python==0.7.2
deprecation==2.1.0
docopt==0.6.2
dkpro-cassis==0.7.0
fastapi==0.74.0
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-resources==5.4.0
jieba==0.42.1
Jinja2==3.0.3
langcodes==3.3.0
lxml==4.7.1
MarkupSafe==2.1.0
more-itertools==8.12.0
murmurhash==1.0.6
numpy==1.22.2
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2-dicts-uk==2.4.1.1.1460299261
pyparsing==3.0.7
pythainlp==3.0.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spacy-pkuseg==0.0.28
srsly==2.4.2
starlette==0.17.1
SudachiPy==0.4.9
thinc==8.0.13
tinydb==4.7.0
toposort==1.7
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
wasabi==0.9.0
watchgod==0.7
websockets==10.1
zipp==3.7.0
pandas==1.4.3
python-dateutil==2.8.2